fixed major performance update

+++ DIY/DIY/Tool/Brush.cs
@@ -36,7 +36,7 @@ namespace DIY.Tool
         {
             DIYProject project = mw.Project;
             Color co = mw.ColorPicker.GetColor();
-            DIYColor c = new DIYColor((int)(Opacity / 100D * 255D), co.R, co.G, co.B);
+            DIYColor c = new DIYColor((byte)(Opacity / 100D * 255D), co.R, co.G, co.B);
 
             Layer lay = project.Layers[project.SelectedLayer];
             if(lay is ImageLayer)
@@ -62,7 +62,7 @@ namespace DIY.Tool
             if (action == null) return;
             DIYProject project = mw.Project;
             Color co = mw.ColorPicker.GetColor();
-            DIYColor c = new DIYColor((int)(Opacity / 100D * 255D), co.R, co.G, co.B);
+            DIYColor c = new DIYColor((byte)(Opacity / 100D * 255D), co.R, co.G, co.B);
 
             Layer lay = project.Layers[project.SelectedLayer];
             if (lay is ImageLayer)
